refactor: remove listeners v2 feature flag and store
