+++ operate/client/src/App/ProcessInstance/BottomPanel/VariablePanel/Listeners/index.tsx
@@ -6,7 +6,18 @@
  * except in compliance with the Camunda License 1.0.
  */
 
+import {useState} from 'react';
 import {observer} from 'mobx-react';
+import type {Job} from '@camunda/camunda-api-zod-schemas/8.8';
+import type {UseInfiniteQueryResult} from '@tanstack/react-query';
+import {Layer} from '@carbon/react';
+
+import type {RequestError} from 'modules/request';
+import {EmptyMessage} from 'modules/components/EmptyMessage';
+import {flowNodeMetaDataStore} from 'modules/stores/flowNodeMetaData';
+import {spaceAndCapitalize} from 'modules/utils/spaceAndCapitalize';
+import {formatDate} from 'modules/utils/date';
+
 import {
   CellContainer,
   Content,
@@ -16,16 +27,6 @@ import {
   EmptyMessageWrapper,
   Stack,
 } from './styled';
-import {spaceAndCapitalize} from 'modules/utils/spaceAndCapitalize';
-import {
-  MAX_LISTENERS_STORED,
-  processInstanceListenersStore,
-} from 'modules/stores/processInstanceListeners';
-import {formatDate} from 'modules/utils/date';
-import {useState} from 'react';
-import {Layer} from '@carbon/react';
-import {EmptyMessage} from 'modules/components/EmptyMessage';
-import {flowNodeMetaDataStore} from 'modules/stores/flowNodeMetaData';
 
 const FilterLabelMapping = {
   'All listeners': 'ALL_LISTENERS',
@@ -40,153 +41,157 @@ type SelectedItem = {
   selectedItem: FilterLabelMappingKeys;
 };
 
-const ROW_HEIGHT = 46;
+type UseJobsResult = UseInfiniteQueryResult<Job[], RequestError>;
+type Props = {
+  jobs: Job[] | undefined;
+  setListenerTypeFilter: React.Dispatch<
+    React.SetStateAction<'EXECUTION_LISTENER' | 'TASK_LISTENER' | undefined>
+  >;
+  fetchNextPage: UseJobsResult['fetchNextPage'];
+  fetchPreviousPage: UseJobsResult['fetchPreviousPage'];
+  hasNextPage: boolean;
+  hasPreviousPage: boolean;
+};
 
-const Listeners: React.FC = observer(() => {
-  const {
+const Listeners: React.FC<Props> = observer(
+  ({
+    jobs,
     setListenerTypeFilter,
-    shouldFetchPreviousListeners,
-    fetchPreviousListeners,
-    shouldFetchNextListeners,
-    fetchNextListeners,
-  } = processInstanceListenersStore;
-  const {listeners, latestFetch} = processInstanceListenersStore.state;
+    fetchNextPage,
+    fetchPreviousPage,
+    hasNextPage,
+    hasPreviousPage,
+  }) => {
+    const {metaData} = flowNodeMetaDataStore.state;
+    const isUserTask = metaData?.instanceMetadata?.flowNodeType === 'USER_TASK';
 
-  const {metaData} = flowNodeMetaDataStore.state;
-  const isUserTask = metaData?.instanceMetadata?.flowNodeType === 'USER_TASK';
+    const [selectedOption, setSelectedOption] =
+      useState<FilterLabelMappingKeys>('All listeners');
 
-  const [selectedOption, setSelectedOption] =
-    useState<FilterLabelMappingKeys>('All listeners');
+    const handleEmptyMessages = () => {
+      if (isUserTask) {
+        if (selectedOption === 'All listeners') {
+          return 'This flow node has no execution listeners nor user task listeners';
+        }
 
-  const handleEmptyMessages = () => {
-    if (isUserTask) {
-      if (selectedOption === 'All listeners') {
-        return 'This flow node has no execution listeners nor user task listeners';
+        if (selectedOption === 'User task listeners') {
+          return 'This flow node has no user task listeners';
+        }
       }
 
-      if (selectedOption === 'User task listeners') {
-        return 'This flow node has no user task listeners';
-      }
-    }
+      return 'This flow node has no execution listeners';
+    };
 
-    return 'This flow node has no execution listeners';
-  };
+    const listeners = jobs ?? [];
 
-  return (
-    <Content>
-      {isUserTask && (
-        <Dropdown
-          id="listenerTypeFilter"
-          data-testid="listener-type-filter"
-          titleText="Listener type"
-          label="All listeners"
-          hideLabel
-          onChange={async ({selectedItem}: SelectedItem) => {
-            if (selectedItem !== selectedOption) {
-              setSelectedOption(selectedItem);
+    return (
+      <Content>
+        {isUserTask && (
+          <Dropdown
+            id="listenerTypeFilter"
+            data-testid="listener-type-filter"
+            titleText="Listener type"
+            label="All listeners"
+            hideLabel
+            onChange={async ({selectedItem}: SelectedItem) => {
+              if (selectedItem !== selectedOption) {
+                setSelectedOption(selectedItem);
 
-              if (FilterLabelMapping[selectedItem] !== 'ALL_LISTENERS') {
-                setListenerTypeFilter(FilterLabelMapping[selectedItem]);
-              } else {
-                setListenerTypeFilter(undefined);
+                if (FilterLabelMapping[selectedItem] !== 'ALL_LISTENERS') {
+                  setListenerTypeFilter(
+                    FilterLabelMapping[selectedItem] as
+                      | 'EXECUTION_LISTENER'
+                      | 'TASK_LISTENER',
+                  );
+                } else {
+                  setListenerTypeFilter(undefined);
+                }
               }
+            }}
+            items={Object.keys(FilterLabelMapping)}
+            size="sm"
+            selectedItem={selectedOption}
+            disabled={
+              selectedOption === 'All listeners' && listeners?.length === 0
             }
-          }}
-          items={Object.keys(FilterLabelMapping)}
-          size="sm"
-          selectedItem={selectedOption}
-          disabled={
-            selectedOption === 'All listeners' && listeners?.length === 0
-          }
-        />
-      )}
-      <Stack as={Layer}>
-        {listeners?.length > 0 ? (
-          <StructuredList
-            dataTestId="listeners-list"
-            headerColumns={[
-              {cellContent: 'Listener type', width: '20%'},
-              {cellContent: 'Listener key', width: '20%'},
-              {cellContent: 'State', width: '10%'},
-              {cellContent: 'Job type', width: '15%'},
-              {cellContent: 'Event', width: '15%'},
-              {cellContent: 'Time', width: '20%'},
-            ]}
-            headerSize="sm"
-            verticalCellPadding="var(--cds-spacing-02)"
-            label="Listeners List"
-            onVerticalScrollStartReach={async (scrollDown) => {
-              if (shouldFetchPreviousListeners() === false) {
-                return;
-              }
-
-              await fetchPreviousListeners();
-
-              if (
-                listeners?.length === MAX_LISTENERS_STORED &&
-                latestFetch?.itemsCount !== 0 &&
-                latestFetch !== null
-              ) {
-                scrollDown(latestFetch.itemsCount * ROW_HEIGHT);
-              }
-            }}
-            onVerticalScrollEndReach={() => {
-              if (shouldFetchNextListeners() === false) {
-                return;
-              }
-
-              fetchNextListeners();
-            }}
-            rows={listeners?.map(
-              ({listenerType, listenerKey, state, jobType, event, time}) => {
-                return {
-                  key: listenerKey,
-                  dataTestId: listenerKey,
-                  columns: [
-                    {
-                      cellContent: (
-                        <CellContainer orientation="horizontal" gap={3}>
-                          {spaceAndCapitalize(listenerType)}
-                          {state === 'FAILED' && <WarningFilled />}
-                        </CellContainer>
-                      ),
-                    },
-                    {cellContent: <CellContainer>{listenerKey}</CellContainer>},
-                    {
-                      cellContent: (
-                        <CellContainer>
-                          {spaceAndCapitalize(state)}
-                        </CellContainer>
-                      ),
-                    },
-                    {
-                      cellContent: <CellContainer>{jobType}</CellContainer>,
-                    },
-                    {
-                      cellContent: (
-                        <CellContainer>
-                          {spaceAndCapitalize(event)}
-                        </CellContainer>
-                      ),
-                    },
-                    {
-                      cellContent: (
-                        <CellContainer>{formatDate(time)}</CellContainer>
-                      ),
-                    },
-                  ],
-                };
-              },
-            )}
           />
-        ) : (
-          <EmptyMessageWrapper>
-            <EmptyMessage message={handleEmptyMessages()} />
-          </EmptyMessageWrapper>
         )}
-      </Stack>
-    </Content>
-  );
-});
+        <Stack as={Layer}>
+          {listeners?.length > 0 ? (
+            <StructuredList
+              dataTestId="listeners-list"
+              headerColumns={[
+                {cellContent: 'Listener type', width: '20%'},
+                {cellContent: 'Listener key', width: '20%'},
+                {cellContent: 'State', width: '10%'},
+                {cellContent: 'Job type', width: '15%'},
+                {cellContent: 'Event', width: '15%'},
+                {cellContent: 'Time', width: '20%'},
+              ]}
+              headerSize="sm"
+              verticalCellPadding="var(--cds-spacing-02)"
+              label="Listeners List"
+              onVerticalScrollStartReach={() => {
+                if (hasPreviousPage) {
+                  fetchPreviousPage();
+                }
+              }}
+              onVerticalScrollEndReach={() => {
+                if (hasNextPage) {
+                  fetchNextPage();
+                }
+              }}
+              rows={listeners?.map(
+                ({kind, jobKey, state, type, listenerEventType, endTime}) => {
+                  return {
+                    key: jobKey,
+                    dataTestId: jobKey,
+                    columns: [
+                      {
+                        cellContent: (
+                          <CellContainer orientation="horizontal" gap={3}>
+                            {spaceAndCapitalize(kind)}
+                            {state === 'FAILED' && <WarningFilled />}
+                          </CellContainer>
+                        ),
+                      },
+                      {cellContent: <CellContainer>{jobKey}</CellContainer>},
+                      {
+                        cellContent: (
+                          <CellContainer>
+                            {spaceAndCapitalize(state)}
+                          </CellContainer>
+                        ),
+                      },
+                      {
+                        cellContent: <CellContainer>{type}</CellContainer>,
+                      },
+                      {
+                        cellContent: (
+                          <CellContainer>
+                            {spaceAndCapitalize(listenerEventType)}
+                          </CellContainer>
+                        ),
+                      },
+                      {
+                        cellContent: (
+                          <CellContainer>{formatDate(endTime)}</CellContainer>
+                        ),
+                      },
+                    ],
+                  };
+                },
+              )}
+            />
+          ) : (
+            <EmptyMessageWrapper>
+              <EmptyMessage message={handleEmptyMessages()} />
+            </EmptyMessageWrapper>
+          )}
+        </Stack>
+      </Content>
+    );
+  },
+);
 
 export {Listeners};
