+++ operate/client/src/App/ProcessInstance/BottomPanel/VariablePanel/index.tsx
@@ -6,7 +6,7 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import React, {useEffect, useMemo, useState} from 'react';
+import React, {useMemo, useState} from 'react';
 import {observer} from 'mobx-react';
 
 import {flowNodeSelectionStore} from 'modules/stores/flowNodeSelection';
@@ -14,15 +14,10 @@ import {TabView} from 'modules/components/TabView';
 import {useProcessInstancePageParams} from '../../useProcessInstancePageParams';
 import {InputOutputMappings} from './InputOutputMappings';
 import {VariablesContent} from './VariablesContent';
-import {Listeners} from './Listeners/v2';
+import {Listeners} from './Listeners';
 import {WarningFilled} from './styled';
 import {useJobs, type JobsFilter} from 'modules/queries/jobs/useJobs';
 import {useIsRootNodeSelected} from 'modules/hooks/flowNodeSelection';
-
-// TODO: Remove when listeners tab is fully migrated to v2
-import {processInstanceListenersStore} from 'modules/stores/processInstanceListeners';
-import {Listeners as ListenersLegacy} from './Listeners';
-import {IS_LISTENERS_TAB_V2} from 'modules/feature-flags';
 import type {ListenerEntity} from 'modules/types/operate';
 
 type Props = {
@@ -41,10 +36,6 @@ const VariablePanel: React.FC<Props> = observer(function VariablePanel({
 
   const [listenerTypeFilter, setListenerTypeFilter] =
     useState<ListenerEntity['listenerType']>();
-
-  // TODO: remove when listeners tab is fully migrated to v2
-  const {listenersFailureCount, fetchListeners, reset} =
-    processInstanceListenersStore;
 
   const jobsFilter = useMemo<JobsFilter | undefined>(() => {
     if (!flowNodeInstanceId && !flowNodeId) {
@@ -69,49 +60,11 @@ const VariablePanel: React.FC<Props> = observer(function VariablePanel({
     hasPreviousPage,
   } = useJobs({
     payload: {filter: jobsFilter},
-    disabled: !IS_LISTENERS_TAB_V2 || !jobsFilter,
+    disabled: !jobsFilter,
     select: (data) => data.pages?.flatMap((page) => page.items),
   });
 
   const hasFailedListeners = jobs?.some(({state}) => state === 'FAILED');
-
-  // TODO: remove when listeners tab is fully migrated to v2
-  useEffect(() => {
-    if (IS_LISTENERS_TAB_V2) {
-      return;
-    }
-    reset();
-  }, [flowNodeId, flowNodeInstanceId, reset]);
-  useEffect(() => {
-    if (IS_LISTENERS_TAB_V2) {
-      return;
-    }
-    if (flowNodeInstanceId) {
-      fetchListeners({
-        fetchType: 'initial',
-        processInstanceId: processInstanceId,
-        payload: {
-          flowNodeInstanceId,
-          ...(listenerTypeFilter && {listenerTypeFilter}),
-        },
-      });
-    } else if (flowNodeId) {
-      fetchListeners({
-        fetchType: 'initial',
-        processInstanceId: processInstanceId,
-        payload: {
-          flowNodeId,
-          ...(listenerTypeFilter && {listenerTypeFilter}),
-        },
-      });
-    }
-  }, [
-    fetchListeners,
-    processInstanceId,
-    flowNodeId,
-    flowNodeInstanceId,
-    listenerTypeFilter,
-  ]);
 
   return (
     <TabView
@@ -145,46 +98,28 @@ const VariablePanel: React.FC<Props> = observer(function VariablePanel({
                 },
               },
             ]),
-        ...(IS_LISTENERS_TAB_V2
-          ? [
-              {
-                id: 'listeners',
-                testId: 'listeners-tab-button',
-                ...(hasFailedListeners && {
-                  labelIcon: <WarningFilled />,
-                }),
-                label: 'Listeners',
-                content: (
-                  <Listeners
-                    jobs={jobs}
-                    setListenerTypeFilter={setListenerTypeFilter}
-                    fetchNextPage={fetchNextPage}
-                    fetchPreviousPage={fetchPreviousPage}
-                    hasNextPage={hasNextPage}
-                    hasPreviousPage={hasPreviousPage}
-                  />
-                ),
-                removePadding: true,
-                onClick: () => {
-                  setListenerTabVisibility(true);
-                },
-              },
-            ]
-          : [
-              {
-                id: 'listeners',
-                testId: 'listeners-tab-button',
-                ...(listenersFailureCount > 0 && {
-                  labelIcon: <WarningFilled />,
-                }),
-                label: 'Listeners',
-                content: <ListenersLegacy />,
-                removePadding: true,
-                onClick: () => {
-                  setListenerTabVisibility(true);
-                },
-              },
-            ]),
+        {
+          id: 'listeners',
+          testId: 'listeners-tab-button',
+          ...(hasFailedListeners && {
+            labelIcon: <WarningFilled />,
+          }),
+          label: 'Listeners',
+          content: (
+            <Listeners
+              jobs={jobs}
+              setListenerTypeFilter={setListenerTypeFilter}
+              fetchNextPage={fetchNextPage}
+              fetchPreviousPage={fetchPreviousPage}
+              hasNextPage={hasNextPage}
+              hasPreviousPage={hasPreviousPage}
+            />
+          ),
+          removePadding: true,
+          onClick: () => {
+            setListenerTabVisibility(true);
+          },
+        },
       ]}
       key={`tabview-${flowNodeId}-${flowNodeInstanceId}`}
     />
