+++ operate/client/src/modules/feature-flags.ts
@@ -7,6 +7,5 @@
  */
 
 const IS_ADD_TOKEN_WITH_ANCESTOR_KEY_SUPPORTED = false;
-const IS_LISTENERS_TAB_V2 = true;
 
-export {IS_ADD_TOKEN_WITH_ANCESTOR_KEY_SUPPORTED, IS_LISTENERS_TAB_V2};
+export {IS_ADD_TOKEN_WITH_ANCESTOR_KEY_SUPPORTED};
